refactor(cli): move check orchestration from main.rs into check.rs (#182)
main.rs was ~1450 lines of arg parsing plus the whole `ry check`
pipeline. Move the command's orchestration — run_check (config/flag
merge, discovery, watch loop), run_check_once, CheckResult,
CheckContext, and their private helpers (init_tracing, flag_set,
demote_non_source_paths, render_diagnostics, report_check_parse_failure,
is_probably_not_r_source, load_user_stubs, sort_and_deduplicate_*,
report_truncation, rescan, sync_stamps) — into check.rs, the module
that already owns the unified diagnostics query. main.rs keeps the
clap surface (Cli/Cmd/CheckArgs/ColorChoice/ConfidenceChoice and the
explain/typeshed/version/completion commands) and dispatch.

Pure move: no logic changes; the moved unit tests moved with the
code (21 tests, all green; only the ColorChoice policy test remains
in main.rs). dump.rs imports the shared discovery helpers from
check:: now. LOC-neutral within the crate.
